data-engineering/zoomcamp/assignments/module-7/pyproject.toml
[project]
name = "module-7-assignment"
version = "0.1.0"
description = "Module 7 - Streaming with Kafka and PyFlink"
requires-python = ">=3.12"
dependencies = [
    "kafka-python>=2.3.0",
    "pandas>=2.2.0",
    "psycopg2-binary>=2.9.11",
    "pyarrow>=19.0.0",
]
